refactor(api): move findTargetProfiles to its bounded context
