--- api/lib/domain/usecases/get-available-target-profiles-for-organization.js
@@ -1,5 +1,0 @@
-const getAvailableTargetProfilesForOrganization = function ({ organizationId, TargetProfileForSpecifierRepository }) {
-  return TargetProfileForSpecifierRepository.availableForOrganization(organizationId);
-};
-
-export { getAvailableTargetProfilesForOrganization };
