+++ api/src/prescription/target-profile/domain/usecases/get-available-target-profiles-for-organization.js
@@ -1,0 +1,5 @@
+const getAvailableTargetProfilesForOrganization = function ({ organizationId, targetProfileForSpecifierRepository }) {
+  return targetProfileForSpecifierRepository.availableForOrganization(organizationId);
+};
+
+export { getAvailableTargetProfilesForOrganization };
