+++ api/src/prescription/target-profile/domain/usecases/index.js
@@ -5,6 +5,9 @@ import * as targetProfileForAdminRepository from '../../../../shared/infrastruct
 import * as learningContentConversionService from '../../../../../lib/domain/services/learning-content/learning-content-conversion-service.js';
 import * as learningContentRepository from '../../../../../lib/infrastructure/repositories/learning-content-repository.js';
 import * as adminMemberRepository from '../../../../shared/infrastructure/repositories/admin-member-repository.js';
+
+import * as targetProfileForSpecifierRepository from '../../infrastructure/repositories/target-profile-for-specifier-repository.js';
+
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 
@@ -13,6 +16,7 @@ const dependencies = {
   learningContentConversionService,
   learningContentRepository,
   adminMemberRepository,
+  targetProfileForSpecifierRepository,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
