+++ api/src/prescription/target-profile/routes.js
@@ -1,5 +1,6 @@
 import * as adminTargetProfileRoutes from './application/admin-target-profile-route.js';
+import * as targetargetProfileRoutes from './application/target-profile-route.js';
 
-const targetProfileRoutes = [adminTargetProfileRoutes];
+const targetProfileRoutes = [adminTargetProfileRoutes, targetargetProfileRoutes];
 
 export { targetProfileRoutes };
