+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -4,9 +4,7 @@ const { map: _map, omit: _omit } = lodash;
 import {
   expect,
   knex,
-  learningContentBuilder,
   databaseBuilder,
-  mockLearningContent,
   generateValidRequestAuthorizationHeader,
   insertUserWithRoleSuperAdmin,
   sinon,
@@ -1099,69 +1097,6 @@ describe('Acceptance | Application | organization-controller', function () {
       expect(response.statusCode).to.equal(200);
       const archivedOrganization = response.result.data.attributes;
       expect(archivedOrganization['archivist-full-name']).to.equal(`${adminUser.firstName} ${adminUser.lastName}`);
-    });
-  });
-
-  describe('GET /api/organizations/{id}/target-profiles', function () {
-    context('when user is authenticated', function () {
-      let user;
-      let linkedOrganization;
-
-      beforeEach(async function () {
-        const learningContent = [
-          {
-            id: 'recArea0',
-            competences: [
-              {
-                id: 'recNv8qhaY887jQb2',
-                index: '1.3',
-                name: 'Traiter des données',
-              },
-            ],
-          },
-        ];
-        const learningContentObjects = learningContentBuilder.fromAreas(learningContent);
-        mockLearningContent(learningContentObjects);
-
-        user = databaseBuilder.factory.buildUser({});
-        linkedOrganization = databaseBuilder.factory.buildOrganization({});
-        databaseBuilder.factory.buildMembership({
-          userId: user.id,
-          organizationId: linkedOrganization.id,
-        });
-
-        await databaseBuilder.commit();
-      });
-
-      it('should return 200', async function () {
-        const options = {
-          method: 'GET',
-          url: `/api/organizations/${linkedOrganization.id}/target-profiles`,
-          headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
-        };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-      });
-    });
-
-    context('when user is not authenticated', function () {
-      it('should return 401', async function () {
-        const options = {
-          method: 'GET',
-          url: '/api/organizations/1/target-profiles',
-          headers: { authorization: 'Bearer mauvais token' },
-        };
-
-        // when
-        const response = await server.inject(options);
-
-        // then
-        expect(response.statusCode).to.equal(401);
-      });
     });
   });
